Move stringlabels specific code to a dedicated file
This functions are specific to the default labels implementation (used to be known as stringlabels).
